adds minBlueCorrection to output min blue correction
see ticket #2701. If hartmann collimate is called with the
minBlueCorrection keyword, the message indicating how many
degrees to move the blue ring will show only the rotation needed
to get the focus into focus tolerance plus a 1 degree buffer.
For instance, if the tolerance is 4 degrees and bres is 4.5,
normally the routing would ask to move the blue ring 9 degrees
(twice the bres). If minBlueCorrection is set, it would request
2 x (4.5 - 4) + 1 = 2 degrees.

Tests are included.