fixes ticket 2705, in which zero moves were being sent to the collimator
See https://trac.sdss.org/ticket/2705 for a detailed explanation.
Because the move is a float, there where cases in which e.g. piston=0.2
but that ended up sending a boss moveColl piston=0. There was a check in
place to avoid that but it was not taking the integer of piston before
checking whether it was zero.